% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glm.R
\name{evaluate_glm}
\alias{evaluate_glm}
\title{Evaluate GLM model}
\usage{
evaluate_glm(model, data)
}
\arguments{
\item{model}{A trained GLM model object.}

\item{data}{A data frame containing hospital data.}
}
\value{
A list containing evaluation metrics (e.g., RMSE, MAE).
}
\description{
This function evaluates the performance of a trained GLM model using evaluation metrics such as RMSE and MAE.
}
\examples{
data <- read.csv(system.file("data", "hospital_data.csv", package = "BestGLM"))
data <- process_data(data)
model <- train_glm(data, formula = LoS ~ comorbidities, family = "poisson")
}
