% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess hospital data}
\usage{
preprocess_data(data)
}
\arguments{
\item{data}{A data frame containing hospital data}
}
\value{
Preprocessed data frame
Preprocess hospital data

Preprocessed data frame
}
\description{
Preprocess hospital data
}
