% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glm.R
\name{select_best_glm}
\alias{select_best_glm}
\title{Select best GLM model}
\usage{
select_best_glm(data, formulas, families)
}
\arguments{
\item{data}{A data frame containing hospital data.}

\item{formulas}{A list of formulas to try.}

\item{families}{A list of distribution families to try (e.g., "poisson", "negbinomial", "gamma").}
}
\value{
The best GLM model object.
}
\description{
This function selects the best GLM model from a list of formulas and distribution families based on evaluation metrics.
}
\examples{
data <- read.csv(system.file("data", "hospital_data.csv", package = "BestGLM"))
data<- process_data(data)
formulas <- list(LoS ~ comorbidities)
families <- c("poisson", "negbinomial", "gamma")
best_model <- select_best_glm(data, formulas, families)
}
