% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glm.R
\name{train_glm}
\alias{train_glm}
\title{Train GLM model}
\usage{
train_glm(data, formula, family)
}
\arguments{
\item{data}{A data frame containing hospital data.}

\item{formula}{A formula specifying the model to be fitted.}

\item{family}{The distribution family for the GLM model (e.g., "poisson", "negbinomial", "gamma").}
}
\value{
A trained GLM model object.
}
\description{
This function trains a Generalized Linear Model (GLM) using the provided formula and hospital data.
}
\examples{
data <- read.csv(system.file("data", "hospital_data.csv", package = "BestGLM"))
data<- process_data(data)
model <- glm(data, formula = LoS ~ comorbidities, family = "poisson")
}
